metrics: handle metrics plugin being nil
We extend the metrics protocol over `nil`—this will allow us the metrics
plugin to be set to `nil`, which will behave equivalent to a disabled
metrics plugin.

This is useful for users who might be trying out goose during
development, or simply wanting to run it without a metrics plugin.