worker: remove default metrics plugin from opts
Fixes #96.

Having a statsd constructor call in a top-level `def` causes the
underlying `clj-statsd` library to spawn a thread pool for agents. This
causes leiningen to get stuck when doing an AOT compilation to an
uberjar, where all top-level defs are evaluated. We should avoid having
side effects in our top-level defines anyway. Let's leave this option to
the client.