Add type hint to `Thread/sleep`
JDK 19 introduced one more single argument implementation of `Thread/sleep` which accepts `Duration`
This made it ambigous for the JVM interop to figure out which `Thread/sleep` implementation to call(the `long` or the `Duration` implementation).
By passing an explicit `long` type hint to `Thread/sleep`, we bypass this problem.

Co-authored-by: Siri PR <siri@nilenso.com>